manager/requirements.txt
Flask
base58
torch
torchvision
pillow
numpy
